backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dj-database-url>=3.0.1",
    "django>=5.2.4",
    "django-extensions>=4.1",
    "django-filter>=25.1",
    "django-polymorphic>=4.1.0",
    "django-rest-polymorphic>=0.1.10",
    "django-silk>=5.4.0",
    "djangorestframework>=3.16.0",
    "drf-nested-routers>=0.94.2",
    "drf-polymorphic>=2.1.1",
    "markdown>=3.8.2",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.1",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
]
